external references
Replace hard-coded bibliography references with ivoatex-style
BibTeX citations.  In accordance with the "Citation Style"
section in the IvoaTeX user note, the bibliography entries are
now in general only cited once, and the standard name (or whatever)
is used for subsequent references, so the text has been somewhat
reworded in places to accommodate this usage.

A couple of items are included in a new local bibliography file
(localrefs.bib).  One of these should maybe be pushed upstream
to the (submodule) ivoatex repository, but I believe ivoatex
is in transit between volute and github at the moment, so I'm
leaving that for now in case the changes get lost.