correct MIME type citation
Reference 4 in the original DataLink-1.0 claimed to be to a document
listed as "IANA, MIME Media Types", but the URL pointed at the MIME
registration for the text/tab-separated-values media type.
I'm pretty certain this is not what was intended.  From the context
(MIME type syntax discussion) I think that RFC2045 is most likely
the intended reference, so I've replaced it by that.