Cleaning up previous section on "Resources".
This mainly takes up observations of caproles
(http://ivoa.net/documents/caproles/); there is no point in having
availability on the resource level, and we'll probably have to revisit
how we want to do VOSI capabilities.

Fortunately, neither of this is relevant for datalink, which is entirely
unconcerned with the Registry, so we can safely remove any material that
will only give us a headache later.

I've also dropped any comments about mutliple datalink capabilities.
It's unclear to me what such a thing might mean.  Let's rack our brains if
someone actually wants to do this.  Differing auth, anyway, would be
on the interface level, not the capability level, and again, let's talk
about then when there's actually services *and* clients doing this.