normalise double-quotes in LaTeX source
Avoid use of the Unicode 0x201D double-quote character in favour of
` and ' in LaTeX source.  In some cases these were the wrong way round,
in some cases not.  You can avoid this commit if you disagree, but
I've found that processing these characters can cause trouble sometimes.

At the same time, replace use of "..." quoting with ``...'' for
consistency and because the LaTeX output looks better.