Clarification of the meaning and use of semantics and content_qualifier.
This introduces a potentially document-breaking change, namely the requirement
that datalink/core concept URIs must be relative (i.e., not include the URI).
I think everyone has always done it like this, and making this guaranteed makes
it a bit simpler to correctly deal with the semantics column (actually,
both implementations I know that use values from the semantics columns
already make that assumption).

This depends on the ivoatex update that comes with PR #70 for citations
to resolve.