remove orphaned example
Delete badly written example that makes no sense where it appears.

This was probably left over as debris from commit 2dd8d569.
So maybe it should be reinstated in what is now section 3.1.1,
but if so please improve it (replace "bla bla bla", improve
formatting etc) as I requested in #59.