adjust exampleURL ID value
I've removed the "%22" (double quote) sequences surrounding the
exampleURL ID value.  They're not illegal, but they look a bit weird
for an ID and it seems to add unnecessary confusion to this example.

Related: this ID value is, according to the text explaining the example,
supposed to be an `<obs_publisher_did value>`; "NGC 6946" doesn't look
like a typical obs_publisher_did.  Does that matter?