fix misusage of some trailing backslashes
Removed usage of trailing backslashes after some \rfc* macros
where they are not followed by whitespace.

The point of the "\xxx\ " idiom is to prevent the whitespace
that follows a macro from being interpreted as simply the macro
name delimiter and thus swallowed.
Writing "\xxx\ " stops that happening by placing a whitespace macro
after the \xxx macro; if you write something like "\rfcoptional\." it
is interpreted as the "\rfcoptional" macro followed by the "\." macro,
which isn't what you want.

An alternative way to ensure that the macro is terminated without
swallowing following whitespace is to write "{\xxx}".