description field is optional in {links} response
Correct the "value required" entry in Table 1 from "yes" to "no"
for the description field.  This is necessary for consistency
with Sec 3.2.5 that says it's a SHOULD not a MUST.

This was actually a transcription error from the DataLink 1.0
PDF to the ivoadoc LaTeX source file (my error), so it should
have been "no" all along, as it effectively was in DataLink 1.0.