Avoid double-quotes outside verbatim
If the double-quote character '"' is used in LaTeX source outside
of the verbatim environment \verb macro, the HTML conversion often
combines it with the following character to produce some weird
unintended output character.  PDF output appears to be unaffected.

I don't know whether that's a bug in TtH, but stop it happening
here in any case by avoiding such usages in the LaTeX source.